chore(frontend): rename non-snake-case modules to use snake-case names
