+++ crates/frontend/src/pages.rs
@@ -1,10 +1,8 @@
-#![allow(non_snake_case)]
-
-pub mod ContextOverride;
-pub mod Dimensions;
-pub mod Experiment;
-pub mod Home;
-pub mod NotFound;
+pub mod context_override;
 pub mod default_config;
+pub mod dimensions;
+pub mod experiment;
 pub mod experiment_list;
 pub mod function;
+pub mod home;
+pub mod not_found;
